requirements.txt
# core
openai
numpy

# plotting
matplotlib
seaborn

# utilities
tqdm
tiktoken
python-dotenv